# src/main/java/com/bank/credit_card/infraestructure/web/api/controller/CardManagementApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bank.credit_card.infraestructure.web.api.controller;

import com.bank.credit_card.infraestructure.web.api.schema.request.ConsumptionRequest;
import com.bank.credit_card.infraestructure.web.api.schema.request.CreatePaymentRequest;
import com.bank.credit_card.infraestructure.web.api.schema.request.InitiateCardRequest;
import com.bank.credit_card.infraestructure.web.api.schema.response.*;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotNull;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.BindingResult;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import java.time.LocalDate;
import java.util.UUID;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-04T15:10:48.324834300-05:00[America/Lima]")
@Validated
@Tag(name = "CardManagement", description = "the CardManagement API")
public interface CardManagementApi {

    /**
     * PUT /CardManagement/{cardId}/Control : controlCard
     *
     * @param cardId (required)
     * @return (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "controlCard",
            summary = "controlCard",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.PUT,
            value = "/CardManagement/{cardId}/Control",
            produces = {"application/json"}
    )
    ResponseEntity<DefaultResponse2xx> controlCard(
            @Parameter(name = "cardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cardId") Long cardId
    );


    /**
     * PUT /CardManagement/{cardId}/Consumption/{ConsumptionId}/Control : controlConsumption
     *
     * @param cardId        (required)
     * @param consumptionId (required)
     * @return (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "controlConsumption",
            summary = "controlConsumption",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.PUT,
            value = "/CardManagement/{cardId}/Consumption/{ConsumptionId}/Control",
            produces = {"application/json"}
    )
    ResponseEntity<DefaultResponse2xx> controlConsumption(
            @Parameter(name = "cardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cardId") Long cardId,
            @Parameter(name = "ConsumptionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("ConsumptionId") UUID consumptionId
    );


    /**
     * PUT /CardManagement/{cardId}/Payment/{paymentId}/Control : controlPayment
     *
     * @param cardId    (required)
     * @param paymentId (required)
     * @return (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "controlPayment",
            summary = "controlPayment",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.PUT,
            value = "/CardManagement/{cardId}/Payment/{paymentId}/Control",
            produces = {"application/json"}
    )
    ResponseEntity<DefaultResponse2xx> controlPayment(
            @Parameter(name = "cardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cardId") Long cardId,
            @Parameter(name = "paymentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("paymentId") UUID paymentId
    );


    /**
     * POST /CardManagement/{cardId}/Payment/Initiate : initiatePayment
     *
     * @param cardId               (required)
     * @param createPaymentRequest (required)
     * @return (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "createPayment",
            summary = "initiatePayment",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = "/CardManagement/{cardId}/Payment/Initiate",
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    ResponseEntity<DefaultResponse2xx> createPayment(
            @Parameter(name = "cardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cardId") Long cardId,
            @Parameter(name = "CreatePaymentRequest", description = "", required = true) @Valid @RequestBody CreatePaymentRequest createPaymentRequest,
            BindingResult bindingResult
    );


    /**
     * PUT /CardManagement/{cardId}/Consumption/{ConsumptionId}/Exchange : exchangeConsumption
     *
     * @param cardId                     (required)
     * @param consumptionId              (required)
     * @param exchangeConsumptionRequest (required)
     * @return (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "exchangeConsumption",
            summary = "exchangeConsumption",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.PUT,
            value = "/CardManagement/{cardId}/Consumption/{ConsumptionId}/Exchange",
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    ResponseEntity<DefaultResponse2xx> exchangeConsumption(
            @Parameter(name = "cardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cardId") Integer cardId,
            @Parameter(name = "ConsumptionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("ConsumptionId") UUID consumptionId,
            @Parameter(name = "ExchangeConsumptionRequest", description = "", required = true) @Valid @RequestBody ExchangeConsumptionRequest exchangeConsumptionRequest,
            BindingResult bindingResult
    );


    /**
     * POST /CardManagement/Initiate : initiateCard
     *
     * @param initiateCardRequest (required)
     * @return (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "initiateCard",
            summary = "initiateCard",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = "/CardManagement/Initiate",
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    ResponseEntity<DefaultResponse2xx> initiateCard(
            @Parameter(name = "InitiateCardRequest", description = "", required = true) @Valid @RequestBody InitiateCardRequest initiateCardRequest,
            BindingResult bindingResult
    );


    /**
     * POST /CardManagement/{cardId}/Consumption/Initiate : initiateConsumption
     *
     * @param cardId             (required)
     * @param consumptionRequest (required)
     * @return (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "initiateConsumption",
            summary = "initiateConsumption",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = "/CardManagement/{cardId}/Consumption/Initiate",
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    ResponseEntity<DefaultResponse2xx> initiateConsumption(
            @Parameter(name = "cardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cardId") Long cardId,
            @Parameter(name = "ConsumptionRequest", description = "", required = true) @Valid @RequestBody ConsumptionRequest consumptionRequest,
            BindingResult bindingResult
    );


    /**
     * GET /CardManagement/{cardId}/Retrieve : retrieveBalance
     *
     * @param cardId (required)
     * @return (status code 200)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "retrieveBalance",
            summary = "retrieveBalance",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "200", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = RetrieveBalance200Response.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/CardManagement/{cardId}/Retrieve",
            produces = {"application/json"}
    )
    ResponseEntity<RetrieveBalance200Response> retrieveBalance(
            @Parameter(name = "cardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cardId") Long cardId
    );


    /**
     * GET /CardManagement/{cardId}/Consumption/{ConsumptionId}/Retrieve : retrieveConsumption
     *
     * @param cardId        (required)
     * @param consumptionId (required)
     * @param dateStart     (required)
     * @param dateEnd       (required)
     * @return (status code 200)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "retrieveConsumption",
            summary = "retrieveConsumption",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "200", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = RetrieveConsumption200Response.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/CardManagement/{cardId}/Consumption/{ConsumptionId}/Retrieve",
            produces = {"application/json"}
    )
    ResponseEntity<RetrieveConsumption200Response> retrieveConsumption(
            @Parameter(name = "cardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cardId") Long cardId,
            @Parameter(name = "ConsumptionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("ConsumptionId") UUID consumptionId,
            @NotNull @Parameter(name = "dateStart", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "dateStart", required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dateStart,
            @NotNull @Parameter(name = "dateEnd", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "dateEnd", required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dateEnd
    );


    /**
     * GET /CardManagement/{cardId}/Payment/{paymentId}/Retrieve : retrievePayment
     *
     * @param cardId    (required)
     * @param paymentId (required)
     * @param dateStart (required)
     * @param dateEnd   (required)
     * @return (status code 200)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "retrievePayment",
            summary = "retrievePayment",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "200", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = RetrievePayment200Response.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse2xx.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/CardManagement/{cardId}/Payment/{paymentId}/Retrieve",
            produces = {"application/json"}
    )
    ResponseEntity<RetrievePayment200Response> retrievePayment(
            @Parameter(name = "cardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cardId") Long cardId,
            @Parameter(name = "paymentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("paymentId") UUID paymentId,
            @NotNull @Parameter(name = "dateStart", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "dateStart", required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dateStart,
            @NotNull @Parameter(name = "dateEnd", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "dateEnd", required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dateEnd
    );

}